refactor(playback): move ChainHandoffLogic + ChainBridgeAction to their own file
